organize code some more to handle transitions
